Remove as much proc macro generated code as possible
Previously, the butchering method didn't have any existence for the type system.
They were used by the procedural macro to generate correct from_borrowed and
from_owned functions, as well as correct return type for each field.

This resulted in a lot of unnecessary code generated at each proc macro call.
This commit tries to reduce it as much as possible.

Each butchering method now has a struct, declared in butcher::methods. They all
implement a common trait, ButcheringMethod, which define the output type and the
owned and borrowed functions (note that this is written by humans, instead of
being generated, as previously done).

As a consequence, the ButcherField trait is much simpler. It is only required
to define a butchering method. The output type and from_owned and from_borrowed
methods are provided thanks to the implementation of ButcheringMethod on the
provided butchering method. The input type has been removed too. It is now a
generic type instead.

Such changes have greatly reduced the amount of generated code in favor of
documented, human-written code. This will make the Butcher derive macro much
less error-prone in the future.