Inclusion of the unbutcher method everywhere
Note: the code currently does not compile.

This commit introduce the unbutcher method for butchering methods, ButcherField
and Butcher. This method allows to convert a ButcheredType into Type.
Unfortunately, this requires us to copy every borrowed data type contained in the
ButcheredType.

The next commits will implement the unbutcher method in the Butcher derive macro.