Add replacement of Self in proc macro code generation
The previous commit added the ReplaceSelf trait. Its goal was to replace each
occurence of Self with any other type. This commit automatically changes the type
in each Field struct so that it does not contain Self anymore.

This should allow to call the derive Butcher proc macro on structs and enums in
which a field's type contains the Self keyword.