Fix implementation of ReplaceSelf
Now - expect a few exceptions - the ReplaceSelf trait implementation is generated
with declarative macros. This allows us to ensure that the generated code is
correct and will make it easier to maintain.

The next few commits will focus on doing roughly the same thing with the
find_generics_in_type and find_lifetimes_in_type functions.